fix: Split off setup-dafny-from-source-action (#28)
The CI was passing on cases of building from source, but only because the testing environment could call the composite sub-action. It turns out that doesn't work when using this action from a different repository. Instead I've extracted out the logic into a separate https://github.com/dafny-lang/setup-dafny-from-source-action repo.